Revert "feat: add tree print for logical plan"
