+++ datafusion/physical-plan/src/display.rs
@@ -24,15 +24,59 @@ use std::fmt::Formatter;
 
 use arrow::datatypes::SchemaRef;
 
-use datafusion_common::display::{
-    tree_render, DisplayAs, DisplayFormatType, GraphvizBuilder, PlanType, StringifiedPlan,
-};
+use datafusion_common::display::{GraphvizBuilder, PlanType, StringifiedPlan};
 use datafusion_expr::display_schema;
 use datafusion_physical_expr::LexOrdering;
 
-use crate::render_tree::{self, RenderTree};
+use crate::render_tree::RenderTree;
 
 use super::{accept, ExecutionPlan, ExecutionPlanVisitor};
+
+/// Options for controlling how each [`ExecutionPlan`] should format itself
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum DisplayFormatType {
+    /// Default, compact format. Example: `FilterExec: c12 < 10.0`
+    ///
+    /// This format is designed to provide a detailed textual description
+    /// of all parts of the plan.
+    Default,
+    /// Verbose, showing all available details.
+    ///
+    /// This form is even more detailed than [`Self::Default`]
+    Verbose,
+    /// TreeRender, displayed in the `tree` explain type.
+    ///
+    /// This format is inspired by DuckDB's explain plans. The information
+    /// presented should be "user friendly", and contain only the most relevant
+    /// information for understanding a plan. It should NOT contain the same level
+    /// of detail information as the  [`Self::Default`] format.
+    ///
+    /// In this mode, each line has one of two formats:
+    ///
+    /// 1. A string without a `=`, which is printed in its own line
+    ///
+    /// 2. A string with a `=` that is treated as a `key=value pair`. Everything
+    ///    before the first `=` is treated as the key, and everything after the
+    ///    first `=` is treated as the value.
+    ///
+    /// For example, if the output of `TreeRender` is this:
+    /// ```text
+    /// Parquet
+    /// partition_sizes=[1]
+    /// ```
+    ///
+    /// It is rendered in the center of a box in the following way:
+    ///
+    /// ```text
+    /// ┌───────────────────────────┐
+    /// │       DataSourceExec      │
+    /// │    --------------------   │
+    /// │    partition_sizes: [1]   │
+    /// │          Parquet          │
+    /// └───────────────────────────┘
+    ///  ```
+    TreeRender,
+}
 
 /// Wraps an `ExecutionPlan` with various methods for formatting
 ///
@@ -224,7 +268,16 @@ impl<'a> DisplayableExecutionPlan<'a> {
     ///
     /// See [`DisplayFormatType::TreeRender`] for more details.
     pub fn tree_render(&self) -> impl fmt::Display + 'a {
-        tree_render(self.inner)
+        struct Wrapper<'a> {
+            plan: &'a dyn ExecutionPlan,
+        }
+        impl fmt::Display for Wrapper<'_> {
+            fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+                let mut visitor = TreeRenderVisitor { f };
+                visitor.visit(self.plan)
+            }
+        }
+        Wrapper { plan: self.inner }
     }
 
     /// Return a single-line summary of the root of the plan
@@ -459,6 +512,491 @@ impl ExecutionPlanVisitor for GraphvizVisitor<'_, '_> {
         self.parents.pop();
         Ok(true)
     }
+}
+
+/// This module implements a tree-like art renderer for execution plans,
+/// based on DuckDB's implementation:
+/// <https://github.com/duckdb/duckdb/blob/main/src/include/duckdb/common/tree_renderer/text_tree_renderer.hpp>
+///
+/// The rendered output looks like this:
+/// ```text
+/// ┌───────────────────────────┐
+/// │    CoalesceBatchesExec    │
+/// └─────────────┬─────────────┘
+/// ┌─────────────┴─────────────┐
+/// │        HashJoinExec       ├──────────────┐
+/// └─────────────┬─────────────┘              │
+/// ┌─────────────┴─────────────┐┌─────────────┴─────────────┐
+/// │       DataSourceExec      ││       DataSourceExec      │
+/// └───────────────────────────┘└───────────────────────────┘
+/// ```
+///
+/// The renderer uses a three-layer approach for each node:
+/// 1. Top layer: renders the top borders and connections
+/// 2. Content layer: renders the node content and vertical connections
+/// 3. Bottom layer: renders the bottom borders and connections
+///
+/// Each node is rendered in a box of fixed width (NODE_RENDER_WIDTH).
+struct TreeRenderVisitor<'a, 'b> {
+    /// Write to this formatter
+    f: &'a mut Formatter<'b>,
+}
+
+impl TreeRenderVisitor<'_, '_> {
+    // Unicode box-drawing characters for creating borders and connections.
+    const LTCORNER: &'static str = "┌"; // Left top corner
+    const RTCORNER: &'static str = "┐"; // Right top corner
+    const LDCORNER: &'static str = "└"; // Left bottom corner
+    const RDCORNER: &'static str = "┘"; // Right bottom corner
+
+    const TMIDDLE: &'static str = "┬"; // Top T-junction (connects down)
+    const LMIDDLE: &'static str = "├"; // Left T-junction (connects right)
+    const DMIDDLE: &'static str = "┴"; // Bottom T-junction (connects up)
+
+    const VERTICAL: &'static str = "│"; // Vertical line
+    const HORIZONTAL: &'static str = "─"; // Horizontal line
+
+    // TODO: Make these variables configurable.
+    const MAXIMUM_RENDER_WIDTH: usize = 240; // Maximum total width of the rendered tree
+    const NODE_RENDER_WIDTH: usize = 29; // Width of each node's box
+    const MAX_EXTRA_LINES: usize = 30; // Maximum number of extra info lines per node
+
+    /// Main entry point for rendering an execution plan as a tree.
+    /// The rendering process happens in three stages for each level of the tree:
+    /// 1. Render top borders and connections
+    /// 2. Render node content and vertical connections
+    /// 3. Render bottom borders and connections
+    pub fn visit(&mut self, plan: &dyn ExecutionPlan) -> Result<(), fmt::Error> {
+        let root = RenderTree::create_tree(plan);
+
+        for y in 0..root.height {
+            // Start by rendering the top layer.
+            self.render_top_layer(&root, y)?;
+            // Now we render the content of the boxes
+            self.render_box_content(&root, y)?;
+            // Render the bottom layer of each of the boxes
+            self.render_bottom_layer(&root, y)?;
+        }
+
+        Ok(())
+    }
+
+    /// Renders the top layer of boxes at the given y-level of the tree.
+    /// This includes:
+    /// - Top corners (┌─┐) for nodes
+    /// - Horizontal connections between nodes
+    /// - Vertical connections to parent nodes
+    fn render_top_layer(
+        &mut self,
+        root: &RenderTree,
+        y: usize,
+    ) -> Result<(), fmt::Error> {
+        for x in 0..root.width {
+            if root.has_node(x, y) {
+                write!(self.f, "{}", Self::LTCORNER)?;
+                write!(
+                    self.f,
+                    "{}",
+                    Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH / 2 - 1)
+                )?;
+                if y == 0 {
+                    // top level node: no node above this one
+                    write!(self.f, "{}", Self::HORIZONTAL)?;
+                } else {
+                    // render connection to node above this one
+                    write!(self.f, "{}", Self::DMIDDLE)?;
+                }
+                write!(
+                    self.f,
+                    "{}",
+                    Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH / 2 - 1)
+                )?;
+                write!(self.f, "{}", Self::RTCORNER)?;
+            } else {
+                let mut has_adjacent_nodes = false;
+                for i in 0..(root.width - x) {
+                    has_adjacent_nodes = has_adjacent_nodes || root.has_node(x + i, y);
+                }
+                if !has_adjacent_nodes {
+                    // There are no nodes to the right side of this position
+                    // no need to fill the empty space
+                    continue;
+                }
+                // there are nodes next to this, fill the space
+                write!(self.f, "{}", &" ".repeat(Self::NODE_RENDER_WIDTH))?;
+            }
+        }
+        writeln!(self.f)?;
+
+        Ok(())
+    }
+
+    /// Renders the content layer of boxes at the given y-level of the tree.
+    /// This includes:
+    /// - Node names and extra information
+    /// - Vertical borders (│) for boxes
+    /// - Vertical connections between nodes
+    fn render_box_content(
+        &mut self,
+        root: &RenderTree,
+        y: usize,
+    ) -> Result<(), fmt::Error> {
+        let mut extra_info: Vec<Vec<String>> = vec![vec![]; root.width];
+        let mut extra_height = 0;
+
+        for (x, extra_info_item) in extra_info.iter_mut().enumerate().take(root.width) {
+            if let Some(node) = root.get_node(x, y) {
+                Self::split_up_extra_info(
+                    &node.extra_text,
+                    extra_info_item,
+                    Self::MAX_EXTRA_LINES,
+                );
+                if extra_info_item.len() > extra_height {
+                    extra_height = extra_info_item.len();
+                }
+            }
+        }
+
+        let halfway_point = extra_height.div_ceil(2);
+
+        // Render the actual node.
+        for render_y in 0..=extra_height {
+            for (x, _) in root.nodes.iter().enumerate().take(root.width) {
+                if x * Self::NODE_RENDER_WIDTH >= Self::MAXIMUM_RENDER_WIDTH {
+                    break;
+                }
+
+                let mut has_adjacent_nodes = false;
+                for i in 0..(root.width - x) {
+                    has_adjacent_nodes = has_adjacent_nodes || root.has_node(x + i, y);
+                }
+
+                if let Some(node) = root.get_node(x, y) {
+                    write!(self.f, "{}", Self::VERTICAL)?;
+
+                    // Rigure out what to render.
+                    let mut render_text = String::new();
+                    if render_y == 0 {
+                        render_text = node.name.clone();
+                    } else if render_y <= extra_info[x].len() {
+                        render_text = extra_info[x][render_y - 1].clone();
+                    }
+
+                    render_text = Self::adjust_text_for_rendering(
+                        &render_text,
+                        Self::NODE_RENDER_WIDTH - 2,
+                    );
+                    write!(self.f, "{render_text}")?;
+
+                    if render_y == halfway_point && node.child_positions.len() > 1 {
+                        write!(self.f, "{}", Self::LMIDDLE)?;
+                    } else {
+                        write!(self.f, "{}", Self::VERTICAL)?;
+                    }
+                } else if render_y == halfway_point {
+                    let has_child_to_the_right =
+                        Self::should_render_whitespace(root, x, y);
+                    if root.has_node(x, y + 1) {
+                        // Node right below this one.
+                        write!(
+                            self.f,
+                            "{}",
+                            Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH / 2)
+                        )?;
+                        if has_child_to_the_right {
+                            write!(self.f, "{}", Self::TMIDDLE)?;
+                            // Have another child to the right, Keep rendering the line.
+                            write!(
+                                self.f,
+                                "{}",
+                                Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH / 2)
+                            )?;
+                        } else {
+                            write!(self.f, "{}", Self::RTCORNER)?;
+                            if has_adjacent_nodes {
+                                // Only a child below this one: fill the reset with spaces.
+                                write!(
+                                    self.f,
+                                    "{}",
+                                    " ".repeat(Self::NODE_RENDER_WIDTH / 2)
+                                )?;
+                            }
+                        }
+                    } else if has_child_to_the_right {
+                        // Child to the right, but no child right below this one: render a full
+                        // line.
+                        write!(
+                            self.f,
+                            "{}",
+                            Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH)
+                        )?;
+                    } else if has_adjacent_nodes {
+                        // Empty spot: render spaces.
+                        write!(self.f, "{}", " ".repeat(Self::NODE_RENDER_WIDTH))?;
+                    }
+                } else if render_y >= halfway_point {
+                    if root.has_node(x, y + 1) {
+                        // Have a node below this empty spot: render a vertical line.
+                        write!(
+                            self.f,
+                            "{}{}",
+                            " ".repeat(Self::NODE_RENDER_WIDTH / 2),
+                            Self::VERTICAL
+                        )?;
+                        if has_adjacent_nodes
+                            || Self::should_render_whitespace(root, x, y)
+                        {
+                            write!(
+                                self.f,
+                                "{}",
+                                " ".repeat(Self::NODE_RENDER_WIDTH / 2)
+                            )?;
+                        }
+                    } else if has_adjacent_nodes
+                        || Self::should_render_whitespace(root, x, y)
+                    {
+                        // Empty spot: render spaces.
+                        write!(self.f, "{}", " ".repeat(Self::NODE_RENDER_WIDTH))?;
+                    }
+                } else if has_adjacent_nodes {
+                    // Empty spot: render spaces.
+                    write!(self.f, "{}", " ".repeat(Self::NODE_RENDER_WIDTH))?;
+                }
+            }
+            writeln!(self.f)?;
+        }
+
+        Ok(())
+    }
+
+    /// Renders the bottom layer of boxes at the given y-level of the tree.
+    /// This includes:
+    /// - Bottom corners (└─┘) for nodes
+    /// - Horizontal connections between nodes
+    /// - Vertical connections to child nodes
+    fn render_bottom_layer(
+        &mut self,
+        root: &RenderTree,
+        y: usize,
+    ) -> Result<(), fmt::Error> {
+        for x in 0..=root.width {
+            if x * Self::NODE_RENDER_WIDTH >= Self::MAXIMUM_RENDER_WIDTH {
+                break;
+            }
+            let mut has_adjacent_nodes = false;
+            for i in 0..(root.width - x) {
+                has_adjacent_nodes = has_adjacent_nodes || root.has_node(x + i, y);
+            }
+            if root.get_node(x, y).is_some() {
+                write!(self.f, "{}", Self::LDCORNER)?;
+                write!(
+                    self.f,
+                    "{}",
+                    Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH / 2 - 1)
+                )?;
+                if root.has_node(x, y + 1) {
+                    // node below this one: connect to that one
+                    write!(self.f, "{}", Self::TMIDDLE)?;
+                } else {
+                    // no node below this one: end the box
+                    write!(self.f, "{}", Self::HORIZONTAL)?;
+                }
+                write!(
+                    self.f,
+                    "{}",
+                    Self::HORIZONTAL.repeat(Self::NODE_RENDER_WIDTH / 2 - 1)
+                )?;
+                write!(self.f, "{}", Self::RDCORNER)?;
+            } else if root.has_node(x, y + 1) {
+                write!(self.f, "{}", &" ".repeat(Self::NODE_RENDER_WIDTH / 2))?;
+                write!(self.f, "{}", Self::VERTICAL)?;
+                if has_adjacent_nodes || Self::should_render_whitespace(root, x, y) {
+                    write!(self.f, "{}", &" ".repeat(Self::NODE_RENDER_WIDTH / 2))?;
+                }
+            } else if has_adjacent_nodes || Self::should_render_whitespace(root, x, y) {
+                write!(self.f, "{}", &" ".repeat(Self::NODE_RENDER_WIDTH))?;
+            }
+        }
+        writeln!(self.f)?;
+
+        Ok(())
+    }
+
+    fn extra_info_separator() -> String {
+        "-".repeat(Self::NODE_RENDER_WIDTH - 9)
+    }
+
+    fn remove_padding(s: &str) -> String {
+        s.trim().to_string()
+    }
+
+    pub fn split_up_extra_info(
+        extra_info: &HashMap<String, String>,
+        result: &mut Vec<String>,
+        max_lines: usize,
+    ) {
+        if extra_info.is_empty() {
+            return;
+        }
+
+        result.push(Self::extra_info_separator());
+
+        let mut requires_padding = false;
+        let mut was_inlined = false;
+
+        // use BTreeMap for repeatable key order
+        let sorted_extra_info: BTreeMap<_, _> = extra_info.iter().collect();
+        for (key, value) in sorted_extra_info {
+            let mut str = Self::remove_padding(value);
+            let mut is_inlined = false;
+            let available_width = Self::NODE_RENDER_WIDTH - 7;
+            let total_size = key.len() + str.len() + 2;
+            let is_multiline = str.contains('\n');
+
+            if str.is_empty() {
+                str = key.to_string();
+            } else if !is_multiline && total_size < available_width {
+                str = format!("{key}: {str}");
+                is_inlined = true;
+            } else {
+                str = format!("{key}:\n{str}");
+            }
+
+            if is_inlined && was_inlined {
+                requires_padding = false;
+            }
+
+            if requires_padding {
+                result.push(String::new());
+            }
+
+            let mut splits: Vec<String> = str.split('\n').map(String::from).collect();
+            if splits.len() > max_lines {
+                let mut truncated_splits = Vec::new();
+                for split in splits.iter().take(max_lines / 2) {
+                    truncated_splits.push(split.clone());
+                }
+                truncated_splits.push("...".to_string());
+                for split in splits.iter().skip(splits.len() - max_lines / 2) {
+                    truncated_splits.push(split.clone());
+                }
+                splits = truncated_splits;
+            }
+            for split in splits {
+                Self::split_string_buffer(&split, result);
+            }
+            if result.len() > max_lines {
+                result.truncate(max_lines);
+                result.push("...".to_string());
+            }
+
+            requires_padding = true;
+            was_inlined = is_inlined;
+        }
+    }
+
+    /// Adjusts text to fit within the specified width by:
+    /// 1. Truncating with ellipsis if too long
+    /// 2. Center-aligning within the available space if shorter
+    fn adjust_text_for_rendering(source: &str, max_render_width: usize) -> String {
+        let render_width = source.chars().count();
+        if render_width > max_render_width {
+            let truncated = &source[..max_render_width - 3];
+            format!("{truncated}...")
+        } else {
+            let total_spaces = max_render_width - render_width;
+            let half_spaces = total_spaces / 2;
+            let extra_left_space = if total_spaces % 2 == 0 { 0 } else { 1 };
+            format!(
+                "{}{}{}",
+                " ".repeat(half_spaces + extra_left_space),
+                source,
+                " ".repeat(half_spaces)
+            )
+        }
+    }
+
+    /// Determines if whitespace should be rendered at a given position.
+    /// This is important for:
+    /// 1. Maintaining proper spacing between sibling nodes
+    /// 2. Ensuring correct alignment of connections between parents and children
+    /// 3. Preserving the tree structure's visual clarity
+    fn should_render_whitespace(root: &RenderTree, x: usize, y: usize) -> bool {
+        let mut found_children = 0;
+
+        for i in (0..=x).rev() {
+            let node = root.get_node(i, y);
+            if root.has_node(i, y + 1) {
+                found_children += 1;
+            }
+            if let Some(node) = node {
+                if node.child_positions.len() > 1
+                    && found_children < node.child_positions.len()
+                {
+                    return true;
+                }
+
+                return false;
+            }
+        }
+
+        false
+    }
+
+    fn split_string_buffer(source: &str, result: &mut Vec<String>) {
+        let mut character_pos = 0;
+        let mut start_pos = 0;
+        let mut render_width = 0;
+        let mut last_possible_split = 0;
+
+        let chars: Vec<char> = source.chars().collect();
+
+        while character_pos < chars.len() {
+            // Treating each char as width 1 for simplification
+            let char_width = 1;
+
+            // Does the next character make us exceed the line length?
+            if render_width + char_width > Self::NODE_RENDER_WIDTH - 2 {
+                if start_pos + 8 > last_possible_split {
+                    // The last character we can split on is one of the first 8 characters of the line
+                    // to not create very small lines we instead split on the current character
+                    last_possible_split = character_pos;
+                }
+
+                result.push(source[start_pos..last_possible_split].to_string());
+                render_width = character_pos - last_possible_split;
+                start_pos = last_possible_split;
+                character_pos = last_possible_split;
+            }
+
+            // check if we can split on this character
+            if Self::can_split_on_this_char(chars[character_pos]) {
+                last_possible_split = character_pos;
+            }
+
+            character_pos += 1;
+            render_width += char_width;
+        }
+
+        if source.len() > start_pos {
+            // append the remainder of the input
+            result.push(source[start_pos..].to_string());
+        }
+    }
+
+    fn can_split_on_this_char(c: char) -> bool {
+        (!c.is_ascii_digit() && !c.is_ascii_uppercase() && !c.is_ascii_lowercase())
+            && c != '_'
+    }
+}
+
+/// Trait for types which could have additional details when formatted in `Verbose` mode
+pub trait DisplayAs {
+    /// Format according to `DisplayFormatType`, used when verbose representation looks
+    /// different from the default one
+    ///
+    /// Should not include a newline
+    fn fmt_as(&self, t: DisplayFormatType, f: &mut Formatter) -> fmt::Result;
 }
 
 /// A new type wrapper to display `T` implementing`DisplayAs` using the `Default` mode
