+++ datafusion/physical-plan/src/lib.rs
@@ -39,7 +39,7 @@ pub use datafusion_physical_expr::{
     expressions, Distribution, Partitioning, PhysicalExpr,
 };
 
-pub use crate::display::{DefaultDisplay, VerboseDisplay};
+pub use crate::display::{DefaultDisplay, DisplayAs, DisplayFormatType, VerboseDisplay};
 pub use crate::execution_plan::{
     collect, collect_partitioned, displayable, execute_input_stream, execute_stream,
     execute_stream_partitioned, get_plan_string, with_new_children_if_necessary,
@@ -51,7 +51,6 @@ pub use crate::stream::EmptyRecordBatchStream;
 pub use crate::topk::TopK;
 pub use crate::visitor::{accept, visit_execution_plan, ExecutionPlanVisitor};
 pub use crate::work_table::WorkTable;
-pub use datafusion_common::display::{DisplayAs, DisplayFormatType};
 pub use spill::spill_manager::SpillManager;
 
 mod ordering;
